Externalize oauth urls
